+++ packages/firestore/src/lite-api/user_data_reader.ts
@@ -801,7 +801,7 @@ export function parseData(
   }
 }
 
-function parseObject(
+export function parseObject(
   obj: Dict<unknown>,
   context: ParseContextImpl
 ): { mapValue: ProtoMapValue } {